Remove Elif - we can't determine here
"Else" is a property of the beginning of the block and how
it was jumped to, while "If" is a property of the end of the
block, specifically if there is a conditional jump.

In between we can't tell here if there are state-changing statements,
such as assignment statements, or whether there are only pure expressions.