Better cross-version support. Better main()
Do not assume the opcode from byte is the same as the current interpreter version.

main() now extracts the code type and from bytecode and calls xdis to
get the bytecode version. This allows us to pass in bytecode files
rather than only a func object. If a func object is passed, we convert
that to its code object.

In order to handle things more generally and support cross python
version bytecode, we rely less on xdis.std and instead get information
from bytecode if that exists and only failing being able to get
this information do we default to the same version that the Python
is currently running.