Finish BLOCK_JOIN first pass...
Things work more or less.

More small changes and tweaks are probably needed.
Like adding an edge in a "FOR" statement, moving psuedo op positions
around, adding more info in assembly listing, and adding a line number
for block 0.

But that is for later!

Also, tidy color names in dotio